Update solution structure
