+++ src/Shark.Fido2.VisualStudio.Template/content/Pages/CredentialDetails.cshtml
@@ -1,5 +1,5 @@
 @page
-@model Shark.Fido2.Sample.Template.Pages.CredentialsDetailsModel
+@model Shark.Fido2.Sample.VisualStudio.Template.Pages.CredentialsDetailsModel
 
 <div class="main">
     <h1>Public Key Credentials</h1>
