+++ src/Shark.Fido2.VisualStudio.Template/content/Pages/Index.cshtml
@@ -1,5 +1,5 @@
 @page
-@model Shark.Fido2.Sample.Template.Pages.IndexModel
+@model Shark.Fido2.Sample.VisualStudio.Template.Pages.IndexModel
 
 <div class="main">
     <h1>Public Key Credentials</h1>
