+++ src/Shark.Fido2.VisualStudio.Template/content/Pages/Index.cshtml.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace Shark.Fido2.Sample.Template.Pages;
+namespace Shark.Fido2.Sample.VisualStudio.Template.Pages;
 
 public class IndexModel : PageModel
 {
