+++ src/Shark.Fido2.VisualStudio.Template/content/wwwroot/js/assertion.js
