+++ src/Shark.Fido2.VisualStudio.Template/content/wwwroot/js/attestation.js
