+++ src/Shark.Fido2.VisualStudio.Template/content/wwwroot/js/site.js
