+++ src/Shark.Fido2.VisualStudio.Template/content/wwwroot/lib/notyf/LICENSE.md
