+++ src/Shark.Fido2.sln
@@ -57,7 +57,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.VisualStudio.Te
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Templates", "Templates", "{9932CECD-B487-4BB5-89C1-62B33F8D0175}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.Sample.Template", "Shark.Fido2.VisualStudio.Template\templates\Shark.Fido2.Sample.Template\Shark.Fido2.Sample.Template.csproj", "{22E41617-7FF5-24A4-36FB-AE38B26C8CA3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shark.Fido2.Sample.VisualStudio.Template", "Shark.Fido2.VisualStudio.Template\content\Shark.Fido2.Sample.VisualStudio.Template.csproj", "{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -157,10 +157,10 @@ Global
 		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{22E41617-7FF5-24A4-36FB-AE38B26C8CA3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{22E41617-7FF5-24A4-36FB-AE38B26C8CA3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{22E41617-7FF5-24A4-36FB-AE38B26C8CA3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{22E41617-7FF5-24A4-36FB-AE38B26C8CA3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -184,7 +184,7 @@ Global
 		{9BED5EF2-38EC-AADA-5FBA-4F9D33D1EC28} = {06FE717D-4AC9-4959-A179-260EE0DBCB27}
 		{D0880235-D8C8-04BD-41F9-3B8EE4C6A911} = {0DBAB265-1DF6-444B-BC5F-7A47857979BB}
 		{A7A2ED4B-6DB4-3A6E-CAD2-0DCB2AC0B9D3} = {9932CECD-B487-4BB5-89C1-62B33F8D0175}
-		{22E41617-7FF5-24A4-36FB-AE38B26C8CA3} = {9932CECD-B487-4BB5-89C1-62B33F8D0175}
+		{7DFCA5E6-F5B2-84A8-15F1-C50051D8B723} = {9932CECD-B487-4BB5-89C1-62B33F8D0175}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6595BEFB-D38C-43C0-BC2E-48B8A742FC45}
