+++ src/Shark.Fido2.slnx
@@ -11,8 +11,8 @@
     <Project Path="Shark.Fido2.SqlServer/Shark.Fido2.SqlServer.csproj" />
   </Folder>
   <Folder Name="/Templates/">
+    <Project Path="Shark.Fido2.VisualStudio.Template/content/Shark.Fido2.Sample.VisualStudio.Template.csproj" />
     <Project Path="Shark.Fido2.VisualStudio.Template/Shark.Fido2.VisualStudio.Template.csproj" />
-    <Project Path="Shark.Fido2.VisualStudio.Template/templates/Shark.Fido2.Sample.Template/Shark.Fido2.Sample.Template.csproj" />
   </Folder>
   <Folder Name="/Tests/">
     <Project Path="../tests/Shark.Fido2.Common.Tests/Shark.Fido2.Common.Tests.csproj" />
